requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
decorator==5.1.1
Django==4.1
django-cpf==0.1.0
djangorestframework==3.13.1
executing==1.0.0
geographiclib==1.52
geopy==2.2.0
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
traitlets==5.3.0
wcwidth==0.2.5
